+++ irs-api/src/main/java/org/eclipse/tractusx/irs/services/IrsItemGraphQueryService.java
@@ -73,6 +73,7 @@ import org.eclipse.tractusx.irs.semanticshub.AspectModel;
 import org.eclipse.tractusx.irs.semanticshub.SemanticsHubFacade;
 import org.eclipse.tractusx.irs.services.validation.SchemaNotFoundException;
 import org.eclipse.tractusx.irs.util.JsonUtil;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.beans.support.MutableSortDefinition;
@@ -109,6 +110,8 @@ public class IrsItemGraphQueryService implements IIrsItemGraphQueryService {
 
     private final String bpdmUrl;
 
+    private final SecurityHelperService securityHelperService;
+
     public IrsItemGraphQueryService(final JobOrchestrator<ItemDataRequest, AASTransferProcess> orchestrator,
             final JobStore jobStore, @Qualifier(JOB_BLOB_PERSISTENCE) final BlobPersistence blobStore,
             final MeterRegistryService meterRegistryService, final SemanticsHubFacade semanticsHubFacade,
@@ -121,11 +124,12 @@ public class IrsItemGraphQueryService implements IIrsItemGraphQueryService {
         this.semanticsHubFacade = semanticsHubFacade;
         this.applicationEventPublisher = applicationEventPublisher;
         this.bpdmUrl = bpdmUrl;
+        this.securityHelperService = new SecurityHelperService();
     }
 
     @Override
     public PageResult getJobsByState(final @NonNull List<JobState> states, final Pageable pageable) {
-        final List<MultiTransferJob> jobs = states.isEmpty() ? jobStore.findAll() : jobStore.findByStates(states);
+        final List<MultiTransferJob> jobs = filterJobs(states);
         final List<JobStatusResult> jobStatusResults = jobs.stream()
                                                            .map(job -> JobStatusResult.builder()
                                                                                       .id(job.getJob().getId())
@@ -138,6 +142,15 @@ public class IrsItemGraphQueryService implements IIrsItemGraphQueryService {
                                                            .toList();
 
         return new PageResult(paginateAndSortResults(pageable, jobStatusResults));
+    }
+
+    private List<MultiTransferJob> filterJobs(final @NotNull List<JobState> states) {
+        final List<MultiTransferJob> jobs = states.isEmpty() ? jobStore.findAll() : jobStore.findByStates(states);
+        if (securityHelperService.isAdmin()) {
+            return jobs;
+        } else {
+            return jobs.stream().filter(multiJob -> multiJob.getJob().getOwner().equals(securityHelperService.getClientIdForViewIrs())).toList();
+        }
     }
 
     @Override
@@ -241,6 +254,11 @@ public class IrsItemGraphQueryService implements IIrsItemGraphQueryService {
         final String idAsString = String.valueOf(jobId);
 
         final Optional<MultiTransferJob> canceled = this.jobStore.cancelJob(idAsString);
+        canceled.ifPresent(cancelledJob -> {
+            if (!securityHelperService.isAdmin() && !cancelledJob.getJob().getOwner().equals(securityHelperService.getClientIdForViewIrs())) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Cannot access and cancel job with id " + jobId + " due to missing privileges.");
+            }
+        });
         canceled.ifPresent(cancelledJob -> applicationEventPublisher.publishEvent(
                 new JobProcessingFinishedEvent(cancelledJob.getJobIdString(), cancelledJob.getJob().getState().name(),
                         cancelledJob.getJobParameter().getCallbackUrl(), cancelledJob.getBatchId())));
@@ -256,6 +274,10 @@ public class IrsItemGraphQueryService implements IIrsItemGraphQueryService {
 
         if (multiTransferJob.isPresent()) {
             final MultiTransferJob multiJob = multiTransferJob.get();
+
+            if (!securityHelperService.isAdmin() && !multiJob.getJob().getOwner().equals(securityHelperService.getClientIdForViewIrs())) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Cannot access job with id " + jobId + " due to missing privileges.");
+            }
 
             final var relationships = new ArrayList<Relationship>();
             final var tombstones = new ArrayList<Tombstone>();