+++ irs-common/src/main/java/org/eclipse/tractusx/irs/common/auth/IrsRoles.java
@@ -20,33 +20,16 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.registryclient.central;
+package org.eclipse.tractusx.irs.common.auth;
 
-import java.util.List;
-
-import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
-import org.eclipse.tractusx.irs.component.assetadministrationshell.IdentifierKeyValuePair;
-import org.eclipse.tractusx.irs.registryclient.decentral.LookupShellsResponse;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
 /**
- * Digital Twin Registry Rest Client
+ * Technical user roles
  */
-public interface DigitalTwinRegistryClient {
-
-    /**
-     * @param aasIdentifier The Asset Administration Shell’s unique id
-     * @return Returns a specific Asset Administration Shell Descriptor
-     */
-    AssetAdministrationShellDescriptor getAssetAdministrationShellDescriptor(String aasIdentifier);
-
-    /**
-     * Returns a list of Asset Administration Shell ids based on Asset identifier key-value-pairs.
-     * Only the Shell ids are returned when all provided key-value pairs match.
-     *
-     * @param assetIds The key-value-pair of an Asset identifier
-     * @return urn uuid string list
-     */
-    LookupShellsResponse getAllAssetAdministrationShellIdsByAssetLink(List<IdentifierKeyValuePair> assetIds);
-
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class IrsRoles {
+    public static final String ADMIN_IRS = "admin_irs";
+    public static final String VIEW_IRS = "view_irs";
 }
-