+++ irs-registry-client/src/main/java/org/eclipse/tractusx/irs/registryclient/central/DigitalTwinRegistryClientLocalStub.java
@@ -25,9 +25,10 @@ package org.eclipse.tractusx.irs.registryclient.central;
 import java.util.Collections;
 import java.util.List;
 
-import org.eclipse.tractusx.irs.data.CxTestDataContainer;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.IdentifierKeyValuePair;
+import org.eclipse.tractusx.irs.data.CxTestDataContainer;
+import org.eclipse.tractusx.irs.registryclient.decentral.LookupShellsResponse;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
@@ -52,7 +53,8 @@ class DigitalTwinRegistryClientLocalStub implements DigitalTwinRegistryClient {
     }
 
     @Override
-    public List<String> getAllAssetAdministrationShellIdsByAssetLink(final List<IdentifierKeyValuePair> assetIds) {
-        return Collections.emptyList();
+    public LookupShellsResponse getAllAssetAdministrationShellIdsByAssetLink(
+            final List<IdentifierKeyValuePair> assetIds) {
+        return LookupShellsResponse.builder().result(Collections.emptyList()).build();
     }
 }