+++ irs-registry-client/src/test/java/org/eclipse/tractusx/irs/registryclient/central/CentralDigitalTwinRegistryServiceTest.java
@@ -24,6 +24,7 @@ package org.eclipse.tractusx.irs.registryclient.central;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -35,12 +36,13 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-import org.eclipse.tractusx.irs.data.CxTestDataContainer;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.Endpoint;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.SubmodelDescriptor;
+import org.eclipse.tractusx.irs.data.CxTestDataContainer;
 import org.eclipse.tractusx.irs.registryclient.DigitalTwinRegistryKey;
 import org.eclipse.tractusx.irs.registryclient.DigitalTwinRegistryService;
+import org.eclipse.tractusx.irs.registryclient.decentral.LookupShellsResponse;
 import org.eclipse.tractusx.irs.registryclient.exceptions.RegistryServiceException;
 import org.eclipse.tractusx.irs.testing.containers.LocalTestDataConfigurationAware;
 import org.junit.jupiter.api.BeforeEach;
@@ -86,7 +88,8 @@ class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAw
         final Endpoint endpoint = shellEndpoints.get(0).getEndpoints().get(0);
 
         assertThat(endpoint.getProtocolInformation().getSubprotocolBody()).contains(existingCatenaXId);
-        assertThat(shellEndpoints.get(0).getSemanticId().getKeys().get(0).getValue()).isEqualTo(singleLevelBomAsBuiltURN);
+        assertThat(shellEndpoints.get(0).getSemanticId().getKeys().get(0).getValue()).isEqualTo(
+                singleLevelBomAsBuiltURN);
         assertThat(shellEndpoints.get(1).getSemanticId().getKeys().get(0).getValue()).isEqualTo(serialPartURN);
     }
 
@@ -96,6 +99,8 @@ class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAw
         final DigitalTwinRegistryKey key = new DigitalTwinRegistryKey(catenaXId, "");
         when(dtRegistryClientMock.getAssetAdministrationShellDescriptor(catenaXId)).thenThrow(
                 new RestClientException("Dummy"));
+        when(dtRegistryClientMock.getAllAssetAdministrationShellIdsByAssetLink(any())).thenReturn(
+                LookupShellsResponse.builder().result(Collections.emptyList()).build());
 
         final var keys = List.of(key);
         assertThatExceptionOfType(RestClientException.class).isThrownBy(
@@ -109,6 +114,8 @@ class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAw
         shellDescriptor.setSubmodelDescriptors(List.of());
 
         when(dtRegistryClientMock.getAssetAdministrationShellDescriptor(catenaXId)).thenReturn(shellDescriptor);
+        when(dtRegistryClientMock.getAllAssetAdministrationShellIdsByAssetLink(any())).thenReturn(
+                LookupShellsResponse.builder().result(Collections.emptyList()).build());
 
         final List<SubmodelDescriptor> submodelEndpoints = dtRegistryFacadeWithMock.fetchShells(
                 List.of(new DigitalTwinRegistryKey(catenaXId, ""))).stream().findFirst().get().getSubmodelDescriptors();
@@ -122,6 +129,8 @@ class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAw
         shellDescriptor.setSubmodelDescriptors(List.of());
 
         when(dtRegistryClientMock.getAssetAdministrationShellDescriptor(catenaXId)).thenReturn(shellDescriptor);
+        when(dtRegistryClientMock.getAllAssetAdministrationShellIdsByAssetLink(any())).thenReturn(
+                LookupShellsResponse.builder().result(Collections.emptyList()).build());
 
         dtRegistryFacadeWithMock.fetchShells(List.of(new DigitalTwinRegistryKey(catenaXId, "")));
 
@@ -137,7 +146,7 @@ class CentralDigitalTwinRegistryServiceTest extends LocalTestDataConfigurationAw
         shellDescriptor.setSubmodelDescriptors(List.of());
 
         when(dtRegistryClientMock.getAllAssetAdministrationShellIdsByAssetLink(anyList())).thenReturn(
-                Collections.singletonList(identification));
+                LookupShellsResponse.builder().result(Collections.singletonList(identification)).build());
         when(dtRegistryClientMock.getAssetAdministrationShellDescriptor(identification)).thenReturn(shellDescriptor);
 
         final List<SubmodelDescriptor> submodelEndpoints = dtRegistryFacadeWithMock.fetchShells(