+++ charts/irs-helm/CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.5.0] - 2023-08-30
+### Added
+- Added configurable values for `digitalTwinRegistry.shellDescriptorTemplate` and `digitalTwinRegistry.lookupShellsTemplate`
+
+### Changed
+- Update IRS version to 3.3.5
+
 ## [6.4.2] - 2023-08-11
 ### Added
 - Added entry to .helmignore to only accept values.yaml